groth16::polynomial

Struct Polynomial

pub struct Polynomial<F>
where
    F: Field,
{ /* private fields */ }

Represents a polynomial over a finite field. Supports scalar and polynomial addition, subtraction, multiplication, division. Can be generated via Lagrangian interpolation over a vector.

Implementations

impl<F: Field> Polynomial<F>

pub fn evaluate_over_srs<T>(&self, srs: &Vec<T>) -> Result<T, Report>
where T: MulAssign<F> + CurveGroup + Debug,

Evaluate the polynomial over an SRS to get the y value without knowing x.

pub fn evaluate(&self, x: &F) -> F

Evaluate polynomial at some point x.

pub fn degree(&self) -> usize

pub fn is_zero(&self) -> bool

pub fn new(vec: Vec<F>) -> Self

Create from a vector of coefficients in ascending degree order x^0, x^1, etc

pub fn interpolate_from_vector(vec: &Vec<F>) -> Self

Find a polynomial by doing Lagrange interpolation over a vector, where the x coordinate is taken to be the index of the element in the vector plus one.

Examples

// Interpolates the set of points [(1,3), (2,10), (3,11)]
interpolate_from_vector(vec![3, 10, 11])